pyproject.toml
[project]
name = "neo4j-agno-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.12",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.24.0",
    "neo4j>=6.0.3",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.12.5",
    "python-docx>=1.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
